+++ recipes/install-collectd.rb
@@ -77,6 +77,29 @@ def install_in_ubuntu
 end
 
 #
+# This function install collectd packages on the Debian OSs
+#
+# The process is :
+# - add the SignalFx PPA
+# - install the collectd package
+#
+
+def install_in_debian
+  package 'apt-transport-https'
+  collectd_ppa_source = node['SignalFx_debian_ppa'][get_debian_os_name]['collectd']['uri']
+  signalfx_collectd_plugin_ppa_source = node['SignalFx_debian_ppa'][get_debian_os_name]['collectd_plugin']['uri']
+  signalfx_keyid = node['SignalFx_debian_ppa']['keyid']
+  execute 'add SignalFx PPA' do
+    command "apt-key adv --keyserver keyserver.ubuntu.com --recv-keys #{signalfx_keyid} && 
+             echo #{collectd_ppa_source} > /etc/apt/sources.list.d/signalfx_collectd.list && 
+             echo #{signalfx_collectd_plugin_ppa_source} > /etc/apt/sources.list.d/signalfx_collectd_plugin.list"
+    action :run
+  end
+  ubuntu_update
+  install_package 'collectd'
+end
+
+#
 # This function is to check the operation system and
 # the version number and install collectd.
 #
@@ -90,6 +113,8 @@ when 'amazon'
   install_in_redhat('amazon', node['platform_version'])
 when 'ubuntu'
   install_in_ubuntu
+when 'debian'
+  install_in_debian
 else
   raise ("Do not support your system #{node['platform']}_#{node['platform_version']}")
 end